pyproject.toml
[tool.poetry]
name = "ukgwa"
version = "0.1.0"
description = "UK Government Web Archive Wagtail site"
authors = ["Torchbox <hello@torchbox.com>"]

[tool.poetry.dependencies]
# key dependencies
python = "~3.13"
django = "~4.2"
wagtail = "~7.1"
# and the rest, alphabetically
dj-database-url = "^3.0.1"
django-basic-auth-ip-whitelist = "~0.8"
django-birdbath = "^2.0.1"
django-crispy-forms = "^2.4"
django-csp = "~3.8"
django-defender = "~0.9.8"
django-extensions = "~4.1"
django-pattern-library = "~1.5"
django-redis = "~6.0"
django-storages = {version = "~1.14", extras = ["boto3"]}
django-xff = "~1.5"
gunicorn = "~23.0"
html5lib = "^1.1"  # used by reading_time template tags
psycopg = "~3.2.10"
scout-apm = "~3.4.0"
sentry-sdk = "~2.39.0"
wagtail-accessibility = "~2.0"
wagtail-storages = "~2.0"
whitenoise = "^6.11.0"

[tool.poetry.group.dev.dependencies]
# testing
coverage = "^7.10.7"
tblib = "^3.1.0"
wagtail-factories = "^4.3.0"

# linting
black = "25.9.0"
detect-secrets = "~1.5.0"
djhtml = "~3.0.9"
flake8 = "7.3.0"
isort = "6.0.1"
pre-commit = "~4.3.0"

# docs
mkdocs = "~1.6.0"
mkdocs-material = "~9.6"
pymdown-extensions = "~10.16"

# utils
django-debug-toolbar = "^6.0.0"
django-upgrade = "^1.29.1"
dslr = "^0.6.0"
fabric = "^3"
honcho = "^2.0"
pudb = "^2025.1.1"
pyupgrade = "^3.21.2"
Werkzeug = "^3.1.3"


[build-system]
requires = ["poetry>=1,<2"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | __pycache__
    | _build
    | build
    | dist
    | docs
    | venv
    | node_modules
  )/
)
'''
